Created Memoize using Map
Memoization is an optimization technique in computer science and programming that involves caching the results of expensive function calls and reusing those cached results when the same inputs occur again. This can greatly improve the performance of functions that are called repeatedly with the same arguments. In JavaScript, you can implement memoization using a higher-order function.

Here's a basic example of a memoize function in JavaScript: